requirements.txt
torch
torchvision
numpy
pandas
opencv-python
pillow
matplotlib
jupyter
